+++ scripts/constants.js
@@ -1,0 +1,2 @@
+export const MODULE_ID = 'daggerheart-fear-macros';
+export const FEAR_SETTING_KEY = 'daggerheart.ResourcesFear';